编写了上下首切换功能（第41集）。

+++ app/src/main/java/com/smart/himalaya/interfaces/IPlayerCallback.java
@@ -52,7 +52,7 @@ public interface IPlayerCallback {
      * @param currentProgress
      * @param total
      */
-    void onProgressChange(long currentProgress, long total);
+    void onProgressChange(int currentProgress, int total);
 
     /**
      * 广告正在加载
